fix: wrap aiohttp sessions with 'async with' to ensure the session is cleaned up
This resolves #14.

To ensure the aiohttp client session is cleaned up properly we can wrap _url_open
with an `async with`, it already accepts a session parameter so we can just pass
it within the control block.

I did some testing with my app and I no longer receive the errors reported in #14.